Naming Convention and Logging Changes.
