--- alerts/alertPlugins.ini
@@ -1,16 +1,0 @@
-[uwsgi]
-chdir = /opt/mozdef/envs/mozdef/alerts/
-uid = mozdef
-mule = alertWorker.py
-pyargv = -c /opt/mozdef/envs/mozdef/alerts/alertWorker.conf
-log-syslog = alertplugins-worker
-log-drain = generated 0 bytes
-socket = /opt/mozdef/envs/mozdef/alerts/alertPlugins.socket
-virtualenv = /opt/mozdef/envs/mozdef/
-master-fifo = /opt/mozdef/envs/mozdef/alerts/alertPlugins.fifo
-procname-master = [m]
-procname-prefix = [alertPlugins]
-never-swap
-pidfile= /var/run/mozdef-alerts/alertPlugins.pid
-vacuum = true
-enable-threads
