+++ alerts/alert_plugins.ini
@@ -1,0 +1,16 @@
+[uwsgi]
+chdir = /opt/mozdef/envs/mozdef/alerts/
+uid = mozdef
+mule = alert_worker.py
+pyargv = -c /opt/mozdef/envs/mozdef/alerts/alert_worker.conf
+log-syslog = alertplugins-worker
+log-drain = generated 0 bytes
+socket = /opt/mozdef/envs/mozdef/alerts/alert_plugins.socket
+virtualenv = /opt/mozdef/envs/mozdef/
+master-fifo = /opt/mozdef/envs/mozdef/alerts/alert_plugins.fifo
+procname-master = [m]
+procname-prefix = [alertplugins]
+never-swap
+pidfile= /var/run/mozdef-alerts/alert_plugins.pid
+vacuum = true
+enable-threads
