+++ alerts/ldap_delete.py
