+++ config/50-mozdef-filter.conf
@@ -1,23 +1,8 @@
 if $programname == 'mozdefbot-worker' then /var/log/mozdef/mozdefbot.log
 if $programname == 'loginput-worker' then /var/log/mozdef/loginput.log
-if $programname == 'infosecsqs-worker' then /var/log/mozdef/infosecsqs.log
 if $programname == 'restapi-worker' then /var/log/mozdef/restapi.log
-if $programname == 'syslog-worker' then /var/log/mozdef/syslog.log
-if $programname == 'nubis-worker' then /var/log/mozdef/nubis.log
-if $programname == 'bro-worker' then /var/log/mozdef/bro.log
-if $programname == 'migsqs-worker' then /var/log/mozdef/migsqs.log
-if $programname == 'parsyssqs-worker' then /var/log/mozdef/parsyssqs.log
-if $programname == 'autoland-worker' then /var/log/mozdef/autoland.log
-if $programname == 'contegix-worker' then /var/log/mozdef/contegix.log
-if $programname == 'deis-worker' then /var/log/mozdef/deis.log
-if $programname == 'releng-worker' then /var/log/mozdef/releng.log
-if $programname == 'fxa-worker' then /var/log/mozdef/fxa.log
-if $programname == 'httpobs-worker' then /var/log/mozdef/httpobs.log
-if $programname == 'riskheatmap-worker' then /var/log/mozdef/riskheatmap.log
-if $programname == 'ssosqs-worker' then /var/log/mozdef/sso.log
-if $programname == 'cloudtrail-worker' then /var/log/mozdef/cloudtrail.log
+if $programname == 'eventtask-worker' then /var/log/mozdef/eventtask.log
 if $programname == 'alertplugins-worker' then /var/log/mozdef/alertplugins.log
-if $programname == 'contegix-auditd-worker' then /var/log/mozdef/contegix-auditd.log
 if $programname == 'mongod.3002' then /var/log/mozdef/mongo/meteor-mongo.log
 if $programname == 'mongod' then /var/log/mozdef/mongo/mongo.log
 if $programname == 'kibana4' then /var/log/mozdef/kibana.log
