+++ config/logrotate-mongod
@@ -1,6 +1,6 @@
 /var/log/mozdef/mongo/*.log
 {
- rotate 17
+ rotate 4
  weekly
  missingok
  notifempty
