+++ mq/esworker_papertrail.conf
