+++ mq/eventtask.ini
@@ -1,0 +1,25 @@
+[uwsgi]
+chdir = /opt/mozdef/envs/mozdef/mq/
+uid = mozdef
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+mule = esworker_eventtask.py
+pyargv = -c /opt/mozdef/envs/mozdef/mq/esworker_eventtask.conf
+log-syslog = eventtask-worker
+log-drain = generated 0 bytes
+socket = /opt/mozdef/envs/mozdef/mq/eventtask.socket
+virtualenv = /opt/mozdef/envs/mozdef/
+procname-master = [m]
+procname-prefix = [eventtask]
+master-fifo = /opt/mozdef/envs/mozdef/mq/eventtask.fifo
+never-swap
+pidfile = /var/run/mozdefmqwSyslog/eventtask.pid
+vacuum = true
+enable-threads
