--- mq/mqwSyslog.ini
@@ -1,25 +1,0 @@
-[uwsgi]
-chdir = /opt/mozdef/envs/mozdef/mq/
-uid = mozdef
-mule = esworker.py
-mule = esworker.py
-mule = esworker.py
-mule = esworker.py
-mule = esworker.py
-mule = esworker.py
-mule = esworker.py
-mule = esworker.py
-mule = esworker.py
-mule = esworker.py
-pyargv = -c /opt/mozdef/envs/mozdef/mq/esworker.conf
-log-syslog = syslog-worker
-log-drain = generated 0 bytes
-socket = /opt/mozdef/envs/mozdef/mq/mqwSyslog.socket
-virtualenv = /opt/mozdef/envs/mozdef/
-procname-master = [m]
-procname-prefix = [mqwSyslog]
-master-fifo = /opt/mozdef/envs/mozdef/mq/mqwSyslog.fifo
-never-swap
-pidfile = /var/run/mozdefmqwSyslog/mqwSyslog.pid
-vacuum = true
-enable-threads
