+++ systemdfiles/alert/mozdefalertplugins.service
@@ -1,5 +1,5 @@
 [Unit]
-Description=uWSGI mozdef Alert Plugins
+Description=uWSGI MozDef Alert Plugins Service
 After=rabbitmq-server.service
 
 [Service]
@@ -9,7 +9,7 @@ ExecStartPre=-/usr/bin/mkdir -p /var/run/mozdef-alerts
 ExecStartPre=/usr/bin/chown -R mozdef:mozdef /var/run/mozdef-alerts
 User=mozdef
 Group=mozdef
-ExecStart=/bin/bash -c 'cd /opt/mozdef/envs/mozdef/alerts; source /opt/mozdef/envs/mozdef/bin/activate; uwsgi --ini alertPlugins.ini'
+ExecStart=/bin/bash -c 'cd /opt/mozdef/envs/mozdef/alerts; source /opt/mozdef/envs/mozdef/bin/activate; uwsgi --ini alert_plugins.ini'
 Restart=always
 KillSignal=SIGQUIT
 Type=notify
@@ -18,4 +18,4 @@ NotifyAccess=all
 
 [Install]
 WantedBy=multi-user.target
-Alias=alertPlugins
+Alias=alertplugins
