+++ systemdfiles/alert/mozdefalerts.service
@@ -1,5 +1,5 @@
 [Unit]
-Description=uWSGI mozdef Alert Service
+Description=uWSGI MozDef Alert Service
 After=rabbitmq-server.service
 
 [Service]
@@ -10,7 +10,7 @@ ExecStartPre=/usr/bin/chown -R mozdef:mozdef /var/run/mozdef-alerts
 PIDFile=/var/run/mozdef-alerts/supervisord.pid
 User=mozdef
 Group=mozdef
-ExecStart=/bin/bash -c 'cd /opt/mozdef/envs/mozdef/alerts; source /opt/mozdef/envs/mozdef/bin/activate; /opt/mozdef/envs/mozdef/bin/supervisord -c /opt/mozdef/envs/mozdef/alerts/supervisord.alerts.ini'
+ExecStart=/bin/bash -c 'cd /opt/mozdef/envs/mozdef/alerts; source /opt/mozdef/envs/mozdef/bin/activate; /opt/mozdef/envs/mozdef/bin/supervisord -c /opt/mozdef/envs/mozdef/alerts/supervisord_alerts.ini'
 Restart=always
 KillSignal=SIGQUIT
 Type=forking
