--- systemdfiles/consumer/mozdefmqwsyslog.service
@@ -1,21 +1,0 @@
-[Unit]
-Description=uWSGI mozdef mqwSyslog
-After=rabbitmq-server.service
-
-[Service]
-# Requires systemd version 211 or newer
-PermissionsStartOnly=true
-ExecStartPre=-/usr/bin/mkdir -p /var/run/mozdefmqwSyslog
-ExecStartPre=/usr/bin/chown -R mozdef:mozdef /var/run/mozdefmqwSyslog
-User=mozdef
-Group=mozdef
-ExecStart=/bin/bash -c 'cd /opt/mozdef/envs/mozdef/mq; source /opt/mozdef/envs/mozdef/bin/activate; uwsgi --ini mqwSyslog.ini'
-Restart=always
-KillSignal=SIGQUIT
-Type=notify
-StandardError=syslog
-NotifyAccess=all
-
-[Install]
-WantedBy=multi-user.target
-Alias=mqwsyslog
