+++ systemdfiles/consumer/mworker-eventtask.service
@@ -1,0 +1,22 @@
+[Unit]
+Description=uWSGI MozDef Eventtask mworker Service
+After=rabbitmq-server.service
+
+[Service]
+# Requires systemd version 211 or newer
+PermissionsStartOnly=true
+ExecStartPre=-/usr/bin/mkdir -p /var/run/mozdef_mq_worker_pids
+ExecStartPre=/usr/bin/chown -R mozdef:mozdef /var/run/mozdef_mq_worker_pids
+User=mozdef
+Group=mozdef
+ExecStart=/bin/bash -c 'cd /opt/mozdef/envs/mozdef/mq; source /opt/mozdef/envs/mozdef/bin/activate; uwsgi --ini eventtask.ini'
+Restart=always
+KillSignal=SIGQUIT
+Type=notify
+StandardError=syslog
+NotifyAccess=all
+
+[Install]
+WantedBy=multi-user.target
+Alias=eventtask
+
